@@ -1,7 +1,10 @@
-apply plugin: 'java'
+plugins {
+    id 'java'
+    id "io.freefair.lombok" version "6.5.1"
+}
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
 
 repositories {
     mavenCentral()
@@ -11,25 +14,25 @@ repositories {
 }
 
 dependencies {
-    compile 'net.portswigger.burp.extender:burp-extender-api:1.7.22'
-    compile 'org.swinglabs:swingx:1.6.1'
-    compile 'com.github.CoreyD97:BurpExtenderUtilities:e800fd2d'
-    compile 'com.google.code.gson:gson:2.8.2'
-    compile 'org.elasticsearch.client:elasticsearch-rest-high-level-client:7.5.2'
-    compile 'org.apache.httpcomponents:httpclient:4.5.6'
-    compile 'org.apache.commons:commons-text:1.7'
-    compile 'org.apache.logging.log4j:log4j-core:2.17.0'
-    // Test on unix system
-     testRuntime files('/home/corey/BurpSuitePro/burpsuite_pro.jar')
-    // Test on Windows System
-//    testRuntime files('C:/Program Files/BurpSuitePro/burpsuite_pro.jar')
+    implementation 'net.portswigger.burp.extensions:montoya-api:2023.5'
+    implementation 'org.swinglabs:swingx:1.6.1'
+    implementation 'com.github.CoreyD97:Burp-Montoya-Utilities:234d21d'
+//    implementation 'org.elasticsearch.client:elasticsearch-rest-high-level-client:7.17.9'
+    implementation 'co.elastic.clients:elasticsearch-java:8.6.2'
+    implementation 'org.apache.httpcomponents:httpclient:4.5.13'
+    implementation 'org.apache.commons:commons-text:1.10.0'
+    implementation 'org.apache.logging.log4j:log4j-core:2.19.0'
+
+    testRuntimeOnly files("${System.properties['user.home']}/BurpSuitePro/burpsuite_pro.jar")
 }
 
-jar{
+jar {
     baseName = project.name
+    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
+
     from {
-        (configurations.compile).collect { it.isDirectory() ? it : zipTree(it) }
-    }{
+        (configurations.runtimeClasspath).collect { it.isDirectory() ? it : zipTree(it) }
+    } {
         exclude "META-INF/*.SF"
         exclude "META-INF/*.DSA"
         exclude "META-INF/*.RSA"
