@@ -1,13 +1,13 @@
 package com.nccgroup.loggerplusplus.exports;
 
-import burp.ICookie;
-import burp.IParameter;
-import burp.IRequestInfo;
-import burp.IResponseInfo;
+import burp.api.montoya.http.message.Cookie;
+import burp.api.montoya.http.message.HttpHeader;
+import burp.api.montoya.http.message.params.HttpParameter;
+import burp.api.montoya.http.message.params.HttpParameterType;
+import burp.api.montoya.http.message.requests.HttpRequest;
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import com.nccgroup.loggerplusplus.LoggerPlusPlus;
 import com.nccgroup.loggerplusplus.logentry.LogEntry;
 import com.nccgroup.loggerplusplus.logentry.LogEntryField;
 
@@ -53,63 +53,57 @@ public class HarSerializer extends TypeAdapter<List<LogEntry>> {
 
             final String pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSZ";
             SimpleDateFormat simpleDateFormat = new SimpleDateFormat(pattern);
-            writer.name("startedDateTime").value(simpleDateFormat.format(logEntry.requestDateTime));
+            writer.name("startedDateTime").value(simpleDateFormat.format(logEntry.getRequestDateTime()));
 
-            long time = logEntry.responseDateTime.getTime() - logEntry.requestDateTime.getTime();
+            long time = logEntry.getResponseDateTime().getTime() - logEntry.getRequestDateTime().getTime();
             time = time < 0 ? 0 : time;
             writer.name("time").value(time);
             writer.name("request").beginObject();
-            writer.name("method").value(logEntry.method);
-            writer.name("url").value(logEntry.url.toString());
-            writer.name("httpVersion").value(logEntry.requestHttpVersion);
+            writer.name("method").value(logEntry.getMethod());
+            writer.name("url").value(logEntry.getUrlString().toString());
+            writer.name("httpVersion").value(logEntry.getRequestHttpVersion());
+            writer.name("origin").value(logEntry.getOrigin());
 
             writer.name("cookies").beginArray();
-            if (logEntry.hasCookieParam) {
-                List<IParameter> cookies = getRequestParametersByType(logEntry.requestResponse.getRequest(),
-                        IParameter.PARAM_COOKIE);
-                for (IParameter cookie : cookies) {
+            if (logEntry.isHasCookieParam()) {
+                List<HttpParameter> cookies = getRequestParametersByType(logEntry.getRequest(), HttpParameterType.COOKIE);
+                for (HttpParameter cookie : cookies) {
                     writer.beginObject();
-                    writer.name("name").value(cookie.getName());
-                    writer.name("value").value(cookie.getValue());
+                    writer.name("name").value(cookie.name());
+                    writer.name("value").value(cookie.value());
                     writer.endObject();
                 }
             }
             writer.endArray(); // end request cookies array
 
             writer.name("headers").beginArray();
-            for (String headerString : logEntry.requestHeaders) {
-                if (headerString.contains(":")) {
-                    writer.beginObject();
-                    String headerArray[] = headerString.split(":", 2);
-                    writer.name("name").value(headerArray[0]);
-                    writer.name("value").value(headerArray[1].trim());
-                    writer.endObject();
-                }
+            for (HttpHeader header : logEntry.getRequestHeaders()) {
+                writer.beginObject();
+                writer.name("name").value(header.name());
+                writer.name("value").value(header.value());
+                writer.endObject();
             }
             writer.endArray(); // end request headers array
 
             writer.name("queryString").beginArray();
-            if (logEntry.url.getQuery() != null) {
-                List<IParameter> queryParams = getRequestParametersByType(logEntry.requestResponse.getRequest(),
-                        IParameter.PARAM_URL);
-                for (IParameter queryParam : queryParams) {
+            if (logEntry.getUrl().getQuery() != null) {
+                for (HttpParameter queryParam : getRequestParametersByType(logEntry.getRequest(), HttpParameterType.URL)) {
                     writer.beginObject();
-                    writer.name("name").value(queryParam.getName());
-                    writer.name("value").value(queryParam.getValue());
+                    writer.name("name").value(queryParam.name());
+                    writer.name("value").value(queryParam.value());
                     writer.endObject();
                 }
             }
             writer.endArray(); // end request queryString array
 
-            if (logEntry.hasBodyParam) {
+            if (logEntry.isHasBodyParam()) {
                 writer.name("postData").beginObject();
-                writer.name("mimeType").value(logEntry.requestContentType);
-                List<IParameter> bodyParams = getRequestBodyParameters(logEntry.requestResponse.getRequest());
+                writer.name("mimeType").value(logEntry.getRequestContentType());
                 writer.name("params").beginArray();
-                for (IParameter bodyParam : bodyParams) {
+                for (HttpParameter bodyParam : getRequestParametersByType(logEntry.getRequest(), HttpParameterType.BODY)) {
                     writer.beginObject();
-                    writer.name("name").value(bodyParam.getName());
-                    writer.name("value").value(bodyParam.getValue());
+                    writer.name("name").value(bodyParam.name());
+                    writer.name("value").value(bodyParam.value());
                     writer.endObject();
                 }
                 writer.endArray(); // end params array
@@ -117,49 +111,46 @@ public class HarSerializer extends TypeAdapter<List<LogEntry>> {
                 writer.endObject(); // end postData object
             }
 
-            writer.name("headersSize").value(logEntry.requestResponse.getRequest().length - logEntry.requestBodyLength);
-            writer.name("bodySize").value(logEntry.requestBodyLength);
+            writer.name("headersSize").value(logEntry.getRequestBytes().length - logEntry.getRequestBodyLength());
+            writer.name("bodySize").value(logEntry.getRequestBodyLength());
 
             writer.endObject(); // end request object
 
             writer.name("response").beginObject();
-            writer.name("status").value(logEntry.responseStatus);
-            writer.name("statusText").value(logEntry.responseStatusText);
-            writer.name("httpVersion").value(logEntry.responseHttpVersion);
+            writer.name("status").value(logEntry.getResponseStatus());
+            writer.name("statusText").value(logEntry.getResponseStatusText());
+            writer.name("httpVersion").value(logEntry.getResponseHttpVersion());
 
             writer.name("cookies").beginArray();
-            if (logEntry.hasSetCookies) {
-                List<ICookie> cookies = getResponseCookies(logEntry.requestResponse.getResponse());
+            if (logEntry.isHasSetCookies()) {
+                List<Cookie> cookies = logEntry.getResponse().cookies();
 
-                for (ICookie cookie : cookies) {
+                for (Cookie cookie : cookies) {
                     writer.beginObject();
-                    writer.name("name").value(cookie.getName());
-                    writer.name("value").value(cookie.getValue());
-                    writer.name("path").value(cookie.getPath());
-                    writer.name("domain").value(cookie.getDomain());
+                    writer.name("name").value(cookie.name());
+                    writer.name("value").value(cookie.value());
+                    writer.name("path").value(cookie.path());
+                    writer.name("domain").value(cookie.domain());
                     writer.endObject();
                 }
             }
             writer.endArray(); // end response cookies array
 
             writer.name("headers").beginArray();
-            if (logEntry.responseHeaders != null) {
-                for (String headerString : logEntry.responseHeaders) {
-                    if (headerString.contains(":")) {
-                        writer.beginObject();
-                        String headerArray[] = headerString.split(":", 2);
-                        writer.name("name").value(headerArray[0]);
-                        writer.name("value").value(headerArray[1].trim());
-                        writer.endObject();
-                    }
+            if (logEntry.getResponseHeaders() != null) {
+                for (HttpHeader header : logEntry.getResponseHeaders()) {
+                    writer.beginObject();
+                    writer.name("name").value(header.name());
+                    writer.name("value").value(header.value());
+                    writer.endObject();
                 }
             }
             writer.endArray(); // end response headers array
 
             writer.name("redirectURL").value(String.valueOf(logEntry.getValueByKey(LogEntryField.REDIRECT_URL)));
-            if (logEntry.requestResponse.getResponse() != null) {
-                writer.name("headersSize").value(logEntry.requestResponse.getResponse().length - logEntry.responseBodyLength);
-                writer.name("bodySize").value(logEntry.responseBodyLength);
+            if (logEntry.getResponseBytes() != null) {
+                writer.name("headersSize").value(logEntry.getResponseBytes().length - logEntry.getResponseBodyLength());
+                writer.name("bodySize").value(logEntry.getResponseBodyLength());
             } else {
                 writer.name("headersSize").value(0);
                 writer.name("bodySize").value(0);
@@ -167,8 +158,8 @@ public class HarSerializer extends TypeAdapter<List<LogEntry>> {
 
 
             writer.name("content").beginObject(); // start content object
-            writer.name("size").value(logEntry.responseBodyLength);
-            writer.name("mimeType").value(logEntry.responseMimeType);
+            writer.name("size").value(logEntry.getResponseBodyLength());
+            writer.name("mimeType").value(logEntry.getResponseContentType());
             writer.name("text").value(String.valueOf(logEntry.getValueByKey(LogEntryField.RESPONSE_BODY)));
             writer.endObject(); //end content object
 
@@ -194,32 +185,17 @@ public class HarSerializer extends TypeAdapter<List<LogEntry>> {
 
     }
 
-    private List<IParameter> getRequestParametersByType(byte[] request, byte paramType) {
-        IRequestInfo tempAnalyzedReq = LoggerPlusPlus.callbacks.getHelpers().analyzeRequest(request);
-        List<IParameter> params = tempAnalyzedReq.getParameters().stream()
-                .filter(iParameter -> iParameter.getType() == paramType).collect(Collectors.toList());
-        return params;
+    private List<HttpParameter> getRequestParametersByType(HttpRequest request, HttpParameterType paramType) {
+        return request.parameters().stream()
+                .filter(iParameter -> iParameter.type().equals(paramType))
+                .collect(Collectors.toList());
     }
 
-    private List<IParameter> getRequestBodyParameters(byte[] request) {
-        IRequestInfo tempAnalyzedReq = LoggerPlusPlus.callbacks.getHelpers().analyzeRequest(request);
-        List<IParameter> params = tempAnalyzedReq.getParameters().stream()
-                .filter(iParameter -> iParameter.getType() != IParameter.PARAM_COOKIE
-                        && iParameter.getType() != IParameter.PARAM_URL)
-                .collect(Collectors.toList());
-        return params;
-    }
 
     @Override
     public List<LogEntry> read(JsonReader reader) throws IOException {
         // TODO Implement HAR Import logic
         return null;
-    }
-
-    private List<ICookie> getResponseCookies(byte[] responseMessage) {
-        IResponseInfo tempAnalyzedResp = LoggerPlusPlus.callbacks.getHelpers().analyzeResponse(responseMessage);
-
-        return tempAnalyzedResp.getCookies();
     }
 
 }
