@@ -1,16 +1,18 @@
 package com.nccgroup.loggerplusplus;
 
-import burp.IContextMenuFactory;
-import burp.IContextMenuInvocation;
-import com.nccgroup.loggerplusplus.filter.colorfilter.ColorFilter;
-import com.nccgroup.loggerplusplus.filter.logfilter.LogFilter;
-import com.nccgroup.loggerplusplus.filter.parser.ParseException;
+import burp.api.montoya.core.Range;
+import burp.api.montoya.http.message.HttpMessage;
+import burp.api.montoya.ui.contextmenu.ContextMenuEvent;
+import burp.api.montoya.ui.contextmenu.ContextMenuItemsProvider;
+import burp.api.montoya.ui.contextmenu.MessageEditorHttpRequestResponse;
+import com.nccgroup.loggerplusplus.filter.colorfilter.TableColorRule;
 import com.nccgroup.loggerplusplus.logentry.LogEntryField;
 import com.nccgroup.loggerplusplus.logview.logtable.LogTable;
 import com.nccgroup.loggerplusplus.util.userinterface.dialog.ColorFilterDialog;
 import org.apache.commons.text.StringEscapeUtils;
 
 import javax.swing.*;
+import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -19,66 +21,66 @@ import java.util.UUID;
 
 import static com.nccgroup.loggerplusplus.util.Globals.PREF_COLOR_FILTERS;
 
-public class LoggerContextMenuFactory implements IContextMenuFactory {
-
-    private final LoggerPlusPlus loggerPlusPlus;
+public class LoggerContextMenuFactory implements ContextMenuItemsProvider {
     
-    public LoggerContextMenuFactory(LoggerPlusPlus loggerPlusPlus){
-        this.loggerPlusPlus = loggerPlusPlus;
+    public LoggerContextMenuFactory(){
     }
-    
+
     @Override
-    public List<JMenuItem> createMenuItems(IContextMenuInvocation invocation) {
-        if(invocation == null) return null;
+    public List<Component> provideMenuItems(ContextMenuEvent event) {
         JMenuItem filterMenu = new JMenu("Logger++");
 
-        if (invocation.getSelectedMessages().length == 0 ||
-                invocation.getSelectionBounds()[0] == invocation.getSelectionBounds()[1]) {
-            return null;
-        }
+        //We're handling a message editor context menu
+        //And we have a selection
+        MessageEditorHttpRequestResponse requestResponse = event.messageEditorRequestResponse().orElseThrow();
+        Range selectedRange = requestResponse.selectionOffsets().orElseThrow();
+        HttpMessage target;
 
         final LogEntryField context;
         final byte[] selectedBytes;
-        switch (invocation.getInvocationContext()){
-            case IContextMenuInvocation.CONTEXT_MESSAGE_EDITOR_REQUEST:
-            case IContextMenuInvocation.CONTEXT_MESSAGE_VIEWER_REQUEST: {
+        switch (event.invocationType()){
+            case MESSAGE_EDITOR_REQUEST:
+            case MESSAGE_VIEWER_REQUEST: {
+                target = requestResponse.requestResponse().request();
                 try {
-                    byte[] msg = invocation.getSelectedMessages()[0].getRequest();
-                    if(LoggerPlusPlus.callbacks.getHelpers().analyzeRequest(msg).getBodyOffset() >= invocation.getSelectionBounds()[0]){
+                    if (selectedRange.startIndexInclusive() <= target.bodyOffset()) {
                         context = LogEntryField.REQUEST_HEADERS;
-                    }else{
+                    } else {
                         context = LogEntryField.REQUEST_BODY;
                     }
-                    selectedBytes = Arrays.copyOfRange(invocation.getSelectedMessages()[0].getRequest(),
-                            invocation.getSelectionBounds()[0],invocation.getSelectionBounds()[1]);
+                    selectedBytes = Arrays.copyOfRange(target.toByteArray().getBytes(), selectedRange.startIndexInclusive(),
+                            selectedRange.endIndexExclusive());
                 }catch (NullPointerException nPException){ return null; }
                 break;
             }
 
-            case IContextMenuInvocation.CONTEXT_MESSAGE_EDITOR_RESPONSE:
-            case IContextMenuInvocation.CONTEXT_MESSAGE_VIEWER_RESPONSE: {
-                try{
-                    byte[] msg = invocation.getSelectedMessages()[0].getResponse();
-                    if(LoggerPlusPlus.callbacks.getHelpers().analyzeRequest(msg).getBodyOffset() >= invocation.getSelectionBounds()[0]){
+            case MESSAGE_EDITOR_RESPONSE:
+            case MESSAGE_VIEWER_RESPONSE: {
+                target = requestResponse.requestResponse().response();
+                try {
+                    if (selectedRange.startIndexInclusive() <= target.bodyOffset()) {
                         context = LogEntryField.RESPONSE_HEADERS;
-                    }else{
+                    } else {
                         context = LogEntryField.RESPONSE_BODY;
                     }
-                    selectedBytes = Arrays.copyOfRange(invocation.getSelectedMessages()[0].getResponse(),
-                            invocation.getSelectionBounds()[0],invocation.getSelectionBounds()[1]);
-                }catch (NullPointerException nPException){ return null; }
+                    selectedBytes = Arrays.copyOfRange(target.toByteArray().getBytes(), selectedRange.startIndexInclusive(),
+                            selectedRange.endIndexExclusive());
+                } catch (NullPointerException nPException) {
+                    return null;
+                }
                 break;
             }
-            default: return null;
+            default:
+                return null;
         }
 
-        final LogTable logTable = loggerPlusPlus.getLogViewController().getLogTableController().getLogTable();
+        final LogTable logTable = LoggerPlusPlus.instance.getLogViewController().getLogTableController().getLogTable();
         String selectedText = StringEscapeUtils.escapeJava(new String(selectedBytes));
 
         JMenuItem useAsFilter = new JMenuItem(new AbstractAction("Use Selection As LogFilter") {
             @Override
             public void actionPerformed(ActionEvent actionEvent) {
-                loggerPlusPlus.getLogViewController().getLogFilterController().setFilter(context.getFullLabel() +
+                LoggerPlusPlus.instance.getLogViewController().getLogFilterController().setFilter(context.getFullLabel() +
                         " CONTAINS \"" + selectedText + "\"");
             }
         });
@@ -90,35 +92,39 @@ public class LoggerContextMenuFactory implements IContextMenuFactory {
             JMenuItem andFilter = new JMenuItem(new AbstractAction("AND") {
                 @Override
                 public void actionPerformed(ActionEvent actionEvent) {
-                    loggerPlusPlus.getLogViewController().getLogFilterController().setFilter(logTable.getCurrentFilter().toString() + " && "
+                    LoggerPlusPlus.instance.getLogViewController().getLogFilterController().setFilter(logTable.getCurrentFilter().toString() + " && "
                             + "" + context.getFullLabel() + " CONTAINS \"" + selectedText + "\"");
                 }
             });
+
+            JMenuItem andNotFilter = new JMenuItem(new AbstractAction("AND NOT") {
+                @Override
+                public void actionPerformed(ActionEvent actionEvent) {
+                    LoggerPlusPlus.instance.getLogViewController().getLogFilterController().setFilter(logTable.getCurrentFilter().toString() + " && !("
+                            + "" + context.getFullLabel() + " CONTAINS \"" + selectedText + "\")");
+                }
+            });
+
             JMenuItem orFilter = new JMenuItem(new AbstractAction("OR") {
                 @Override
                 public void actionPerformed(ActionEvent actionEvent) {
-                    loggerPlusPlus.getLogViewController().getLogFilterController().setFilter(logTable.getCurrentFilter().toString() + " || "
+                    LoggerPlusPlus.instance.getLogViewController().getLogFilterController().setFilter(logTable.getCurrentFilter().toString() + " || "
                             + context.getFullLabel() + " CONTAINS \"" + selectedText + "\"");
                 }
             });
             addToCurrentFilter.add(andFilter);
+            addToCurrentFilter.add(andNotFilter);
             addToCurrentFilter.add(orFilter);
             filterMenu.add(addToCurrentFilter);
         }
 
-        JMenuItem colorFilterItem = new JMenuItem(new AbstractAction("Set Selection as Color LogFilter") {
+        JMenuItem colorFilterItem = new JMenuItem(new AbstractAction("Set Selection as Color Filter") {
             @Override
             public void actionPerformed(ActionEvent actionEvent) {
-                try {
-                    ColorFilter colorFilter = new ColorFilter();
-                    colorFilter.setFilter(new LogFilter(loggerPlusPlus.getLibraryController(),
-                            context.getFullLabel() + " CONTAINS \"" + selectedText + "\""));
-                    HashMap<UUID,ColorFilter> colorFilters = loggerPlusPlus.getPreferencesController().getPreferences().getSetting(PREF_COLOR_FILTERS);
-                    colorFilters.put(colorFilter.getUUID(), colorFilter);
-                    new ColorFilterDialog(loggerPlusPlus.getLibraryController()).setVisible(true);
-                } catch (ParseException e) {
-                    return;
-                }
+                TableColorRule tableColorRule = new TableColorRule("New Filter", context.getFullLabel() + " CONTAINS \"" + selectedText + "\"");
+                HashMap<UUID, TableColorRule> colorFilters = LoggerPlusPlus.instance.getPreferencesController().getPreferences().getSetting(PREF_COLOR_FILTERS);
+                colorFilters.put(tableColorRule.getUuid(), tableColorRule);
+                new ColorFilterDialog(LoggerPlusPlus.instance.getLibraryController()).setVisible(true);
             }
         });
         filterMenu.add(colorFilterItem);
