@@ -5,7 +5,7 @@ import com.coreyd97.BurpExtenderUtilities.ComponentGroup;
 import com.coreyd97.BurpExtenderUtilities.PanelBuilder;
 import com.coreyd97.BurpExtenderUtilities.Preferences;
 import com.nccgroup.loggerplusplus.LoggerPlusPlus;
-import com.nccgroup.loggerplusplus.filter.logfilter.LogFilter;
+import com.nccgroup.loggerplusplus.filter.logfilter.LogTableFilter;
 import com.nccgroup.loggerplusplus.filter.parser.ParseException;
 import com.nccgroup.loggerplusplus.filterlibrary.FilterLibraryController;
 import com.nccgroup.loggerplusplus.logentry.LogEntryField;
@@ -221,10 +221,8 @@ public class ElasticExporterConfigDialog extends JDialog {
                 String logFilter = preferences.getSetting(PREF_ELASTIC_FILTER);
 
                 if (!StringUtils.isBlank(logFilter)) {
-                    FilterLibraryController libraryController = elasticExporter.getExportController()
-                            .getLoggerPlusPlus().getLibraryController();
                     try {
-                        new LogFilter(libraryController, logFilter);
+                        new LogTableFilter(logFilter);
                     } catch (ParseException ex) {
                         JOptionPane.showMessageDialog(ElasticExporterConfigDialog.this,
                                 "Cannot save Elastic Exporter configuration. The chosen log filter is invalid: \n" +
