@@ -2,6 +2,7 @@ package com.nccgroup.loggerplusplus.exports;
 
 import com.coreyd97.BurpExtenderUtilities.Alignment;
 import com.coreyd97.BurpExtenderUtilities.PanelBuilder;
+import com.nccgroup.loggerplusplus.LoggerPlusPlus;
 import com.nccgroup.loggerplusplus.logentry.LogEntry;
 
 import javax.swing.*;
@@ -15,7 +16,7 @@ public class JSONExporterControlPanel extends JPanel {
 
         JButton manualSaveButton = new JButton("Export as JSON");
         manualSaveButton.addActionListener(actionEvent -> {
-            final List<LogEntry> entries = jsonExporter.getExportController().getLoggerPlusPlus().getLogEntries();
+            final List<LogEntry> entries = LoggerPlusPlus.instance.getLogEntries();
             jsonExporter.exportEntries(entries);
         });
 
