@@ -1,35 +1,49 @@
 /* Generated By:JJTree&JavaCC: Do not edit this line. FilterParserTokenManager.java */
 package com.nccgroup.loggerplusplus.filter.parser;
+import com.nccgroup.loggerplusplus.filter.ComparisonOperator;
+import com.nccgroup.loggerplusplus.filter.LogicalOperator;
+import com.nccgroup.loggerplusplus.filterlibrary.FilterLibraryController;
+import java.io.StringReader;
+import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.LinkedHashSet;
+import java.util.HashSet;
+import java.util.HashMap;
+import java.util.Set;
+import java.util.Date;
+import java.util.ArrayList;
+import java.util.regex.Pattern;
+import com.nccgroup.loggerplusplus.logentry.LogEntryField;
+import com.nccgroup.loggerplusplus.logentry.FieldGroup;
+import static com.nccgroup.loggerplusplus.logentry.LogEntryField.getFieldsInGroup;
+import com.nccgroup.loggerplusplus.logview.processor.LogProcessor;
 
-/**
- * Token Manager.
- */
-public class FilterParserTokenManager implements FilterParserConstants {
+/** Token Manager. */
+public class FilterParserTokenManager implements FilterParserConstants
+{
 
-   /** Debug output. */
-   public java.io.PrintStream debugStream = System.out;
-
-   /** Set debug output. */
-   public void setDebugStream(java.io.PrintStream ds) {
-      debugStream = ds;
-   }
-
-   private int jjStopAtPos(int pos, int kind) {
-      jjmatchedKind = kind;
-      jjmatchedPos = pos;
-      return pos + 1;
-   }
-
-   private int jjMoveStringLiteralDfa0_0() {
-      switch (curChar) {
-         case 32:
-            jjmatchedKind = 1;
-            return jjMoveNfa_0(0, 0);
-         case 33:
-            return jjMoveStringLiteralDfa1_0(0x8L);
-         case 34:
-            jjmatchedKind = 29;
-            return jjMoveNfa_0(0, 0);
+  /** Debug output. */
+  public  java.io.PrintStream debugStream = System.out;
+  /** Set debug output. */
+  public  void setDebugStream(java.io.PrintStream ds) { debugStream = ds; }
+private int jjStopAtPos(int pos, int kind)
+{
+   jjmatchedKind = kind;
+   jjmatchedPos = pos;
+   return pos + 1;
+}
+private int jjMoveStringLiteralDfa0_0()
+{
+   switch(curChar)
+   {
+      case 32:
+         jjmatchedKind = 1;
+         return jjMoveNfa_0(0, 0);
+      case 33:
+         return jjMoveStringLiteralDfa1_0(0x8L);
+      case 34:
+         jjmatchedKind = 29;
+         return jjMoveNfa_0(0, 0);
       case 35:
          jjmatchedKind = 27;
          return jjMoveNfa_0(0, 0);
@@ -66,27 +80,28 @@ public class FilterParserTokenManager implements FilterParserConstants {
       case 93:
          jjmatchedKind = 25;
          return jjMoveNfa_0(0, 0);
-         case 99:
-            return jjMoveStringLiteralDfa1_0(0x800L);
-         case 105:
-            return jjMoveStringLiteralDfa1_0(0x1000L);
-         case 109:
-            return jjMoveStringLiteralDfa1_0(0x40000L);
-         default:
-            return jjMoveNfa_0(0, 0);
-      }
-   }
-
-   private int jjMoveStringLiteralDfa1_0(long active0) {
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
+      case 99:
+         return jjMoveStringLiteralDfa1_0(0x800L);
+      case 105:
+         return jjMoveStringLiteralDfa1_0(0x1000L);
+      case 109:
+         return jjMoveStringLiteralDfa1_0(0x40000L);
+      default :
          return jjMoveNfa_0(0, 0);
-      }
-      switch (curChar) {
-         case 61:
-            if ((active0 & 0x8L) != 0L) {
-               jjmatchedKind = 3;
+   }
+}
+private int jjMoveStringLiteralDfa1_0(long active0)
+{
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 0);
+   }
+   switch(curChar)
+   {
+      case 61:
+         if ((active0 & 0x8L) != 0L)
+         {
+            jjmatchedKind = 3;
             jjmatchedPos = 1;
          }
          else if ((active0 & 0x40L) != 0L)
@@ -120,117 +135,117 @@ public class FilterParserTokenManager implements FilterParserConstants {
             jjmatchedPos = 1;
          }
          break;
-         case 111:
-            return jjMoveStringLiteralDfa2_0(active0, 0x800L);
-         default:
-            break;
-      }
-      return jjMoveNfa_0(0, 1);
+      case 111:
+         return jjMoveStringLiteralDfa2_0(active0, 0x800L);
+      default :
+         break;
    }
-
-   private int jjMoveStringLiteralDfa2_0(long old0, long active0) {
-      if (((active0 &= old0)) == 0L)
-         return jjMoveNfa_0(0, 1);
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return jjMoveNfa_0(0, 1);
-      }
-      switch (curChar) {
-         case 78:
-            return jjMoveStringLiteralDfa3_0(active0, 0x800L);
+   return jjMoveNfa_0(0, 1);
+}
+private int jjMoveStringLiteralDfa2_0(long old0, long active0)
+{
+   if (((active0 &= old0)) == 0L)
+      return jjMoveNfa_0(0, 1);
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 1);
+   }
+   switch(curChar)
+   {
+      case 78:
+         return jjMoveStringLiteralDfa3_0(active0, 0x800L);
       case 84:
          return jjMoveStringLiteralDfa3_0(active0, 0x40000L);
-         case 110:
-            return jjMoveStringLiteralDfa3_0(active0, 0x800L);
-         case 116:
-            return jjMoveStringLiteralDfa3_0(active0, 0x40000L);
-         default:
-            break;
-      }
-      return jjMoveNfa_0(0, 2);
+      case 110:
+         return jjMoveStringLiteralDfa3_0(active0, 0x800L);
+      case 116:
+         return jjMoveStringLiteralDfa3_0(active0, 0x40000L);
+      default :
+         break;
    }
-
-   private int jjMoveStringLiteralDfa3_0(long old0, long active0) {
-      if (((active0 &= old0)) == 0L)
-         return jjMoveNfa_0(0, 2);
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return jjMoveNfa_0(0, 2);
-      }
-      switch (curChar) {
-         case 67:
-            return jjMoveStringLiteralDfa4_0(active0, 0x40000L);
+   return jjMoveNfa_0(0, 2);
+}
+private int jjMoveStringLiteralDfa3_0(long old0, long active0)
+{
+   if (((active0 &= old0)) == 0L)
+      return jjMoveNfa_0(0, 2);
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 2);
+   }
+   switch(curChar)
+   {
+      case 67:
+         return jjMoveStringLiteralDfa4_0(active0, 0x40000L);
       case 84:
          return jjMoveStringLiteralDfa4_0(active0, 0x800L);
-         case 99:
-            return jjMoveStringLiteralDfa4_0(active0, 0x40000L);
-         case 116:
-            return jjMoveStringLiteralDfa4_0(active0, 0x800L);
-         default:
-            break;
-      }
-      return jjMoveNfa_0(0, 3);
+      case 99:
+         return jjMoveStringLiteralDfa4_0(active0, 0x40000L);
+      case 116:
+         return jjMoveStringLiteralDfa4_0(active0, 0x800L);
+      default :
+         break;
    }
-
-   private int jjMoveStringLiteralDfa4_0(long old0, long active0) {
-      if (((active0 &= old0)) == 0L)
-         return jjMoveNfa_0(0, 3);
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return jjMoveNfa_0(0, 3);
-      }
-      switch (curChar) {
-         case 65:
-            return jjMoveStringLiteralDfa5_0(active0, 0x800L);
+   return jjMoveNfa_0(0, 3);
+}
+private int jjMoveStringLiteralDfa4_0(long old0, long active0)
+{
+   if (((active0 &= old0)) == 0L)
+      return jjMoveNfa_0(0, 3);
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 3);
+   }
+   switch(curChar)
+   {
+      case 65:
+         return jjMoveStringLiteralDfa5_0(active0, 0x800L);
       case 72:
          return jjMoveStringLiteralDfa5_0(active0, 0x40000L);
-         case 97:
-            return jjMoveStringLiteralDfa5_0(active0, 0x800L);
-         case 104:
-            return jjMoveStringLiteralDfa5_0(active0, 0x40000L);
-         default:
-            break;
-      }
-      return jjMoveNfa_0(0, 4);
+      case 97:
+         return jjMoveStringLiteralDfa5_0(active0, 0x800L);
+      case 104:
+         return jjMoveStringLiteralDfa5_0(active0, 0x40000L);
+      default :
+         break;
    }
-
-   private int jjMoveStringLiteralDfa5_0(long old0, long active0) {
-      if (((active0 &= old0)) == 0L)
-         return jjMoveNfa_0(0, 4);
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return jjMoveNfa_0(0, 4);
-      }
-      switch (curChar) {
-         case 69:
-            return jjMoveStringLiteralDfa6_0(active0, 0x40000L);
+   return jjMoveNfa_0(0, 4);
+}
+private int jjMoveStringLiteralDfa5_0(long old0, long active0)
+{
+   if (((active0 &= old0)) == 0L)
+      return jjMoveNfa_0(0, 4);
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 4);
+   }
+   switch(curChar)
+   {
+      case 69:
+         return jjMoveStringLiteralDfa6_0(active0, 0x40000L);
       case 73:
          return jjMoveStringLiteralDfa6_0(active0, 0x800L);
-         case 101:
-            return jjMoveStringLiteralDfa6_0(active0, 0x40000L);
-         case 105:
-            return jjMoveStringLiteralDfa6_0(active0, 0x800L);
-         default:
-            break;
-      }
-      return jjMoveNfa_0(0, 5);
+      case 101:
+         return jjMoveStringLiteralDfa6_0(active0, 0x40000L);
+      case 105:
+         return jjMoveStringLiteralDfa6_0(active0, 0x800L);
+      default :
+         break;
    }
-
-   private int jjMoveStringLiteralDfa6_0(long old0, long active0) {
-      if (((active0 &= old0)) == 0L)
-         return jjMoveNfa_0(0, 5);
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return jjMoveNfa_0(0, 5);
-      }
-      switch (curChar) {
-         case 78:
-            return jjMoveStringLiteralDfa7_0(active0, 0x800L);
+   return jjMoveNfa_0(0, 5);
+}
+private int jjMoveStringLiteralDfa6_0(long old0, long active0)
+{
+   if (((active0 &= old0)) == 0L)
+      return jjMoveNfa_0(0, 5);
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 5);
+   }
+   switch(curChar)
+   {
+      case 78:
+         return jjMoveStringLiteralDfa7_0(active0, 0x800L);
       case 83:
          if ((active0 & 0x40000L) != 0L)
          {
@@ -241,28 +256,29 @@ public class FilterParserTokenManager implements FilterParserConstants {
       case 110:
          return jjMoveStringLiteralDfa7_0(active0, 0x800L);
       case 115:
-         if ((active0 & 0x40000L) != 0L) {
+         if ((active0 & 0x40000L) != 0L)
+         {
             jjmatchedKind = 18;
             jjmatchedPos = 6;
          }
          break;
-         default:
-            break;
-      }
-      return jjMoveNfa_0(0, 6);
+      default :
+         break;
    }
-
-   private int jjMoveStringLiteralDfa7_0(long old0, long active0) {
-      if (((active0 &= old0)) == 0L)
-         return jjMoveNfa_0(0, 6);
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return jjMoveNfa_0(0, 6);
-      }
-      switch (curChar) {
-         case 83:
-            if ((active0 & 0x800L) != 0L)
+   return jjMoveNfa_0(0, 6);
+}
+private int jjMoveStringLiteralDfa7_0(long old0, long active0)
+{
+   if (((active0 &= old0)) == 0L)
+      return jjMoveNfa_0(0, 6);
+   try { curChar = input_stream.readChar(); }
+   catch(java.io.IOException e) {
+   return jjMoveNfa_0(0, 6);
+   }
+   switch(curChar)
+   {
+      case 83:
+         if ((active0 & 0x800L) != 0L)
          {
             jjmatchedKind = 11;
             jjmatchedPos = 7;
@@ -611,14 +627,15 @@ private int jjMoveNfa_0(int startState, int curPos)
       }
       else
       {
-         int hiByte = (int) (curChar >> 8);
+         int hiByte = (int)(curChar >> 8);
          int i1 = hiByte >> 6;
          long l1 = 1L << (hiByte & 077);
          int i2 = (curChar & 0xff) >> 6;
          long l2 = 1L << (curChar & 077);
          do
          {
-            switch(jjstateSet[--i]) {
+            switch(jjstateSet[--i])
+            {
                case 32:
                case 33:
                   if (jjCanMove_0(hiByte, i1, i2, l1, l2))
@@ -629,8 +646,7 @@ private int jjMoveNfa_0(int startState, int curPos)
                   if (jjCanMove_0(hiByte, i1, i2, l1, l2))
                      jjCheckNAddStates(3, 5);
                   break;
-               default:
-                  break;
+               default : break;
             }
          } while(i != startsAt);
       }
@@ -654,39 +670,41 @@ private int jjMoveNfa_0(int startState, int curPos)
    if (curPos < toRet)
       for (i = toRet - Math.min(curPos, seenUpto); i-- > 0; )
          try { curChar = input_stream.readChar(); }
-         catch(java.io.IOException e) { throw new Error("Internal Error : Please send a bug report.");
-         }
+         catch(java.io.IOException e) { throw new Error("Internal Error : Please send a bug report."); }
 
-   if (jjmatchedPos < strPos) {
+   if (jjmatchedPos < strPos)
+   {
       jjmatchedKind = strKind;
       jjmatchedPos = strPos;
-   } else if (jjmatchedPos == strPos && jjmatchedKind > strKind)
+   }
+   else if (jjmatchedPos == strPos && jjmatchedKind > strKind)
       jjmatchedKind = strKind;
 
    return toRet;
 }
-
-   private final int jjStopStringLiteralDfa_1(int pos, long active0) {
-      switch (pos) {
-         default:
-            return -1;
-      }
+private final int jjStopStringLiteralDfa_1(int pos, long active0)
+{
+   switch (pos)
+   {
+      default :
+         return -1;
    }
-
-   private final int jjStartNfa_1(int pos, long active0) {
-      return jjMoveNfa_1(jjStopStringLiteralDfa_1(pos, active0), pos + 1);
+}
+private final int jjStartNfa_1(int pos, long active0)
+{
+   return jjMoveNfa_1(jjStopStringLiteralDfa_1(pos, active0), pos + 1);
+}
+private int jjMoveStringLiteralDfa0_1()
+{
+   switch(curChar)
+   {
+      case 39:
+         return jjStopAtPos(0, 31);
+      default :
+         return jjMoveNfa_1(3, 0);
    }
-
-   private int jjMoveStringLiteralDfa0_1() {
-      switch (curChar) {
-         case 39:
-            return jjStopAtPos(0, 31);
-         default:
-            return jjMoveNfa_1(3, 0);
-      }
-   }
-
-   private int jjMoveNfa_1(int startState, int curPos)
+}
+private int jjMoveNfa_1(int startState, int curPos)
 {
    int startsAt = 0;
    jjnewStateCnt = 3;
@@ -760,14 +778,15 @@ private int jjMoveNfa_0(int startState, int curPos)
       }
       else
       {
-         int hiByte = (int) (curChar >> 8);
+         int hiByte = (int)(curChar >> 8);
          int i1 = hiByte >> 6;
          long l1 = 1L << (hiByte & 077);
          int i2 = (curChar & 0xff) >> 6;
          long l2 = 1L << (curChar & 077);
          do
          {
-            switch(jjstateSet[--i]) {
+            switch(jjstateSet[--i])
+            {
                case 3:
                case 2:
                case 1:
@@ -777,8 +796,7 @@ private int jjMoveNfa_0(int startState, int curPos)
                      kind = 30;
                   jjCheckNAddTwoStates(0, 2);
                   break;
-               default:
-                  break;
+               default : break;
             }
          } while(i != startsAt);
       }
@@ -791,48 +809,50 @@ private int jjMoveNfa_0(int startState, int curPos)
       ++curPos;
       if ((i = jjnewStateCnt) == (startsAt = 3 - (jjnewStateCnt = startsAt)))
          return curPos;
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return curPos;
-      }
+      try { curChar = input_stream.readChar(); }
+      catch(java.io.IOException e) { return curPos; }
    }
 }
-
-   private final int jjStopStringLiteralDfa_2(int pos, long active0) {
-      switch (pos) {
-         default:
-            return -1;
-      }
+private final int jjStopStringLiteralDfa_2(int pos, long active0)
+{
+   switch (pos)
+   {
+      default :
+         return -1;
    }
-
-   private final int jjStartNfa_2(int pos, long active0) {
-      return jjMoveNfa_2(jjStopStringLiteralDfa_2(pos, active0), pos + 1);
+}
+private final int jjStartNfa_2(int pos, long active0)
+{
+   return jjMoveNfa_2(jjStopStringLiteralDfa_2(pos, active0), pos + 1);
+}
+private int jjMoveStringLiteralDfa0_2()
+{
+   switch(curChar)
+   {
+      case 34:
+         return jjStopAtPos(0, 33);
+      default :
+         return jjMoveNfa_2(3, 0);
    }
-
-   private int jjMoveStringLiteralDfa0_2() {
-      switch (curChar) {
-         case 34:
-            return jjStopAtPos(0, 33);
-         default:
-            return jjMoveNfa_2(3, 0);
-      }
-   }
-
-   private int jjMoveNfa_2(int startState, int curPos)
+}
+private int jjMoveNfa_2(int startState, int curPos)
 {
    int startsAt = 0;
    jjnewStateCnt = 3;
    int i = 1;
    jjstateSet[0] = startState;
    int kind = 0x7fffffff;
-   for (;;) {
+   for (;;)
+   {
       if (++jjround == 0x7fffffff)
          ReInitRounds();
-      if (curChar < 64) {
+      if (curChar < 64)
+      {
          long l = 1L << curChar;
-         do {
-            switch(jjstateSet[--i]) {
+         do
+         {
+            switch(jjstateSet[--i])
+            {
                case 3:
                case 2:
                   if ((0xfffffffbffffffffL & l) == 0L)
@@ -849,16 +869,22 @@ private int jjMoveNfa_0(int startState, int curPos)
                default : break;
             }
          } while(i != startsAt);
-      } else if (curChar < 128) {
+      }
+      else if (curChar < 128)
+      {
          long l = 1L << (curChar & 077);
-         do {
-            switch(jjstateSet[--i]) {
+         do
+         {
+            switch(jjstateSet[--i])
+            {
                case 3:
-                  if ((0xffffffffefffffffL & l) != 0L) {
+                  if ((0xffffffffefffffffL & l) != 0L)
+                  {
                      if (kind > 32)
                         kind = 32;
                      jjCheckNAddTwoStates(0, 2);
-                  } else if (curChar == 92)
+                  }
+                  else if (curChar == 92)
                      jjstateSet[jjnewStateCnt++] = 1;
                   break;
                case 0:
@@ -880,14 +906,18 @@ private int jjMoveNfa_0(int startState, int curPos)
                default : break;
             }
          } while(i != startsAt);
-      } else {
-         int hiByte = (int) (curChar >> 8);
+      }
+      else
+      {
+         int hiByte = (int)(curChar >> 8);
          int i1 = hiByte >> 6;
          long l1 = 1L << (hiByte & 077);
          int i2 = (curChar & 0xff) >> 6;
          long l2 = 1L << (curChar & 077);
-         do {
-            switch(jjstateSet[--i]) {
+         do
+         {
+            switch(jjstateSet[--i])
+            {
                case 3:
                case 2:
                case 1:
@@ -897,12 +927,12 @@ private int jjMoveNfa_0(int startState, int curPos)
                      kind = 32;
                   jjCheckNAddTwoStates(0, 2);
                   break;
-               default:
-                  break;
+               default : break;
             }
          } while(i != startsAt);
       }
-      if (kind != 0x7fffffff) {
+      if (kind != 0x7fffffff)
+      {
          jjmatchedKind = kind;
          jjmatchedPos = curPos;
          kind = 0x7fffffff;
@@ -910,124 +940,108 @@ private int jjMoveNfa_0(int startState, int curPos)
       ++curPos;
       if ((i = jjnewStateCnt) == (startsAt = 3 - (jjnewStateCnt = startsAt)))
          return curPos;
-      try {
-         curChar = input_stream.readChar();
-      } catch (java.io.IOException e) {
-         return curPos;
-      }
+      try { curChar = input_stream.readChar(); }
+      catch(java.io.IOException e) { return curPos; }
+   }
+}
+static final int[] jjnextStates = {
+   31, 33, 36, 40, 42, 43, 
+};
+private static final boolean jjCanMove_0(int hiByte, int i1, int i2, long l1, long l2)
+{
+   switch(hiByte)
+   {
+      case 0:
+         return ((jjbitVec2[i2] & l2) != 0L);
+      default :
+         if ((jjbitVec0[i1] & l1) != 0L)
+            return true;
+         return false;
    }
 }
 
-   static final int[] jjnextStates = {
-           31, 33, 36, 40, 42, 43,
-   };
+/** Token literal values. */
+public static final String[] jjstrLiteralImages = {
+"", null, null, "\41\75", "\76", "\74", "\76\75", "\74\75", null, null, null, 
+null, null, null, null, null, null, null, null, "\50", "\51", null, "\56", null, 
+"\133", "\135", "\54", "\43", "\47", "\42", null, "\47", null, "\42", null, };
 
-   private static final boolean jjCanMove_0(int hiByte, int i1, int i2, long l1, long l2) {
-      switch (hiByte) {
-         case 0:
-            return ((jjbitVec2[i2] & l2) != 0L);
-         default:
-            if ((jjbitVec0[i1] & l1) != 0L)
-               return true;
-            return false;
-      }
-   }
+/** Lexer state names. */
+public static final String[] lexStateNames = {
+   "DEFAULT",
+   "SINGLE_QUOTED_STRING",
+   "DOUBLE_QUOTED_STRING",
+};
 
-   /**
-    * Token literal values.
-    */
-   public static final String[] jjstrLiteralImages = {
-           "", null, null, "\41\75", "\76", "\74", "\76\75", "\74\75", null, null, null,
-           null, null, null, null, null, null, null, null, "\50", "\51", null, "\56", null,
-           "\133", "\135", "\54", "\43", "\47", "\42", null, "\47", null, "\42", null,};
+/** Lex State array. */
+public static final int[] jjnewLexState = {
+   -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 
+   -1, -1, -1, 1, 2, -1, 0, -1, 0, -1, 
+};
+static final long[] jjtoToken = {
+   0x7fffffffdL, 
+};
+static final long[] jjtoSkip = {
+   0x2L, 
+};
+protected SimpleCharStream input_stream;
+private final int[] jjrounds = new int[49];
+private final int[] jjstateSet = new int[98];
+protected char curChar;
+/** Constructor. */
+public FilterParserTokenManager(SimpleCharStream stream){
+   if (SimpleCharStream.staticFlag)
+      throw new Error("ERROR: Cannot use a static CharStream class with a non-static lexical analyzer.");
+   input_stream = stream;
+}
 
-   /**
-    * Lexer state names.
-    */
-   public static final String[] lexStateNames = {
-           "DEFAULT",
-           "SINGLE_QUOTED_STRING",
-           "DOUBLE_QUOTED_STRING",
-   };
+/** Constructor. */
+public FilterParserTokenManager(SimpleCharStream stream, int lexState){
+   this(stream);
+   SwitchTo(lexState);
+}
 
-   /**
-    * Lex State array.
-    */
-   public static final int[] jjnewLexState = {
-           -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-           -1, -1, -1, 1, 2, -1, 0, -1, 0, -1,
-   };
-   static final long[] jjtoToken = {
-           0x7fffffffdL,
-   };
-   static final long[] jjtoSkip = {
-           0x2L,
-   };
-   protected SimpleCharStream input_stream;
-   private final int[] jjrounds = new int[49];
-   private final int[] jjstateSet = new int[98];
-   protected char curChar;
+/** Reinitialise parser. */
+public void ReInit(SimpleCharStream stream)
+{
+   jjmatchedPos = jjnewStateCnt = 0;
+   curLexState = defaultLexState;
+   input_stream = stream;
+   ReInitRounds();
+}
+private void ReInitRounds()
+{
+   int i;
+   jjround = 0x80000001;
+   for (i = 49; i-- > 0;)
+      jjrounds[i] = 0x80000000;
+}
 
-   /**
-    * Constructor.
-    */
-   public FilterParserTokenManager(SimpleCharStream stream) {
-      if (SimpleCharStream.staticFlag)
-         throw new Error("ERROR: Cannot use a static CharStream class with a non-static lexical analyzer.");
-      input_stream = stream;
-   }
+/** Reinitialise parser. */
+public void ReInit(SimpleCharStream stream, int lexState)
+{
+   ReInit(stream);
+   SwitchTo(lexState);
+}
 
-   /**
-    * Constructor.
-    */
-   public FilterParserTokenManager(SimpleCharStream stream, int lexState) {
-      this(stream);
-      SwitchTo(lexState);
-   }
+/** Switch to specified lex state. */
+public void SwitchTo(int lexState)
+{
+   if (lexState >= 3 || lexState < 0)
+      throw new TokenMgrError("Error: Ignoring invalid lexical state : " + lexState + ". State unchanged.", TokenMgrError.INVALID_LEXICAL_STATE);
+   else
+      curLexState = lexState;
+}
 
-   /**
-    * Reinitialise parser.
-    */
-   public void ReInit(SimpleCharStream stream) {
-      jjmatchedPos = jjnewStateCnt = 0;
-      curLexState = defaultLexState;
-      input_stream = stream;
-      ReInitRounds();
-   }
-
-   private void ReInitRounds() {
-      int i;
-      jjround = 0x80000001;
-      for (i = 49; i-- > 0; )
-         jjrounds[i] = 0x80000000;
-   }
-
-   /**
-    * Reinitialise parser.
-    */
-   public void ReInit(SimpleCharStream stream, int lexState) {
-      ReInit(stream);
-      SwitchTo(lexState);
-   }
-
-   /**
-    * Switch to specified lex state.
-    */
-   public void SwitchTo(int lexState) {
-      if (lexState >= 3 || lexState < 0)
-         throw new TokenMgrError("Error: Ignoring invalid lexical state : " + lexState + ". State unchanged.", TokenMgrError.INVALID_LEXICAL_STATE);
-      else
-         curLexState = lexState;
-   }
-
-   protected Token jjFillToken() {
-      final Token t;
-      final String curTokenImage;
-      final int beginLine;
-      final int endLine;
-      final int beginColumn;
-      final int endColumn;
-      String im = jjstrLiteralImages[jjmatchedKind];
+protected Token jjFillToken()
+{
+   final Token t;
+   final String curTokenImage;
+   final int beginLine;
+   final int endLine;
+   final int beginColumn;
+   final int endColumn;
+   String im = jjstrLiteralImages[jjmatchedKind];
    curTokenImage = (im == null) ? input_stream.GetImage() : im;
    beginLine = input_stream.getBeginLine();
    beginColumn = input_stream.getBeginColumn();
@@ -1056,11 +1070,14 @@ public Token getNextToken()
   Token matchedToken;
   int curPos = 0;
 
-   EOFLoop:
-   for (; ; ) {
-      try {
-         curChar = input_stream.BeginToken();
-      } catch (java.io.IOException e)
+  EOFLoop :
+  for (;;)
+  {
+   try
+   {
+      curChar = input_stream.BeginToken();
+   }
+   catch(java.io.IOException e)
    {
       jjmatchedKind = 0;
       matchedToken = jjFillToken();
