@@ -1,39 +1,35 @@
 package com.nccgroup.loggerplusplus.filterlibrary;
 
 import com.coreyd97.BurpExtenderUtilities.Preferences;
-import com.nccgroup.loggerplusplus.LoggerPlusPlus;
-import com.nccgroup.loggerplusplus.filter.colorfilter.ColorFilter;
+import com.nccgroup.loggerplusplus.filter.FilterExpression;
+import com.nccgroup.loggerplusplus.filter.colorfilter.TableColorRule;
 import com.nccgroup.loggerplusplus.filter.colorfilter.ColorFilterListener;
-import com.nccgroup.loggerplusplus.filter.logfilter.LogFilter;
+import com.nccgroup.loggerplusplus.filter.logfilter.LogTableFilter;
 import com.nccgroup.loggerplusplus.filter.savedfilter.SavedFilter;
 import com.nccgroup.loggerplusplus.filter.tag.Tag;
 import com.nccgroup.loggerplusplus.filter.tag.TagListener;
 import com.nccgroup.loggerplusplus.preferences.PreferencesController;
 import com.nccgroup.loggerplusplus.util.Globals;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import lombok.extern.log4j.Log4j2;
 
 import java.awt.*;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.UUID;
 
+@Log4j2
 public class FilterLibraryController {
 
-    private final LoggerPlusPlus loggerPlusPlus;
     private final Preferences preferences;
     private final FilterLibraryPanel panel;
     private final ArrayList<FilterLibraryListener> listeners;
     private final ArrayList<SavedFilter> savedFilters;
-    private final HashMap<UUID, ColorFilter> colorFilters;
+    private final HashMap<UUID, TableColorRule> colorFilters;
     private final ArrayList<ColorFilterListener> colorFilterListeners;
     private final HashMap<UUID, Tag> tagFilters;
     private final ArrayList<TagListener> tagListeners;
 
-    Logger logger = LogManager.getLogger(this);
-
-    public FilterLibraryController(LoggerPlusPlus loggerPlusPlus, PreferencesController preferencesController) {
-        this.loggerPlusPlus = loggerPlusPlus;
+    public FilterLibraryController(PreferencesController preferencesController) {
         this.preferences = preferencesController.getPreferences();
         this.listeners = new ArrayList<>();
         this.colorFilterListeners = new ArrayList<>();
@@ -44,15 +40,11 @@ public class FilterLibraryController {
         this.panel = new FilterLibraryPanel(this);
     }
 
-    public LoggerPlusPlus getLoggerPlusPlus() {
-        return loggerPlusPlus;
-    }
-
     public FilterLibraryPanel getFilterLibraryPanel() {
         return panel;
     }
 
-    public ArrayList<SavedFilter> getSavedFilters(){
+    public ArrayList<SavedFilter> getFilterSnippets(){
         return this.savedFilters;
     }
 
@@ -100,52 +92,52 @@ public class FilterLibraryController {
         this.listeners.remove(listener);
     }
 
-    public HashMap<UUID, ColorFilter> getColorFilters() {
+    public HashMap<UUID, TableColorRule> getColorFilters() {
         return colorFilters;
     }
 
-    public void addColorFilter(String title, LogFilter filter){
+    public void addColorFilter(String title, FilterExpression filter){
         this.addColorFilter(title, filter, Color.BLACK, Color.WHITE);
     }
 
-    public void addColorFilter(String title, LogFilter filter, Color foreground, Color background){
-        this.addColorFilter(new ColorFilter(title, filter, foreground, background));
+    public void addColorFilter(String title, FilterExpression filter, Color foreground, Color background){
+        this.addColorFilter(new TableColorRule(title, filter));
     }
 
-    public void addColorFilter(ColorFilter colorFilter){
-        this.colorFilters.put(colorFilter.getUUID(), colorFilter);
+    public void addColorFilter(TableColorRule tableColorRule){
+        this.colorFilters.put(tableColorRule.getUuid(), tableColorRule);
 
         for (ColorFilterListener colorFilterListener : this.colorFilterListeners) {
             try {
-                colorFilterListener.onColorFilterAdd(colorFilter);
+                colorFilterListener.onColorFilterAdd(tableColorRule);
             } catch (Exception e) {
-                logger.error(e);
+                log.error(e);
             }
         }
         saveColorFilters();
     }
 
-    public void removeColorFilter(ColorFilter colorFilter){
+    public void removeColorFilter(TableColorRule tableColorRule){
         synchronized (this.colorFilters){
-            this.colorFilters.remove(colorFilter.getUUID());
+            this.colorFilters.remove(tableColorRule.getUuid());
         }
         for (ColorFilterListener listener : this.colorFilterListeners) {
             try{
-                listener.onColorFilterRemove(colorFilter);
+                listener.onColorFilterRemove(tableColorRule);
             }catch (Exception e){
-                logger.error(e);
+                log.error(e);
             }
         }
         saveColorFilters();
     }
 
     //Called when a filter is modified.
-    public void updateColorFilter(ColorFilter colorFilter){
+    public void updateColorFilter(TableColorRule tableColorRule){
         for (ColorFilterListener listener : this.colorFilterListeners) {
             try{
-                listener.onColorFilterChange(colorFilter);
+                listener.onColorFilterChange(tableColorRule);
             }catch (Exception e){
-                logger.error(e);
+                log.error(e);
             }
         }
         saveColorFilters();
@@ -168,13 +160,13 @@ public class FilterLibraryController {
     }
 
     public void addTag(Tag tag) {
-        this.tagFilters.put(tag.getUUID(), tag);
+        this.tagFilters.put(tag.getUuid(), tag);
 
         for (TagListener listener : this.tagListeners) {
             try {
                 listener.onTagAdd(tag);
             } catch (Exception error) {
-                logger.error(error);
+                log.error(error);
             }
         }
         saveTags();
@@ -182,13 +174,13 @@ public class FilterLibraryController {
 
     public void removeTag(Tag tag) {
         synchronized (this.tagFilters) {
-            this.tagFilters.remove(tag.getUUID());
+            this.tagFilters.remove(tag.getUuid());
         }
         for (TagListener listener : this.tagListeners) {
             try {
                 listener.onTagRemove(tag);
             } catch (Exception error) {
-                logger.error(error);
+                log.error(error);
             }
         }
         saveTags();
@@ -200,7 +192,7 @@ public class FilterLibraryController {
             try {
                 listener.onTagChange(tag);
             } catch (Exception e) {
-                logger.error(e);
+                log.error(e);
             }
         }
         saveTags();
@@ -218,4 +210,18 @@ public class FilterLibraryController {
         this.tagListeners.remove(listener);
     }
 
+    public void propagateChangesToSnippetUsers(SavedFilter savedFilter) {
+        String snippet = savedFilter.getName();
+        for (TableColorRule tableColorRule : this.getColorFilters().values()) {
+            if(tableColorRule.getFilterExpression().getSnippetDependencies().contains(snippet)){
+                updateColorFilter(tableColorRule);
+            }
+        }
+
+        for (Tag tag : this.getTags().values()) {
+            if(tag.getFilterExpression().getSnippetDependencies().contains(snippet)){
+                updateTag(tag);
+            }
+        }
+    }
 }
