@@ -1,7 +1,8 @@
 package com.nccgroup.loggerplusplus.grepper;
 
-import burp.IHttpRequestResponse;
-import burp.IHttpRequestResponseWithMarkers;
+import burp.api.montoya.core.Marker;
+import burp.api.montoya.core.Range;
+import burp.api.montoya.http.message.HttpRequestResponse;
 import com.coreyd97.BurpExtenderUtilities.Preferences;
 import com.nccgroup.loggerplusplus.LoggerPlusPlus;
 import com.nccgroup.loggerplusplus.logentry.LogEntry;
@@ -20,7 +21,6 @@ import java.util.regex.Pattern;
 
 public class GrepperController {
 
-    private final LoggerPlusPlus loggerPlusPlus;
     private final LogTableController logTableController;
     private final Preferences preferences;
     private final GrepperPanel grepPanel;
@@ -29,17 +29,12 @@ public class GrepperController {
 
     private ExecutorService searchExecutor;
 
-    public GrepperController(LoggerPlusPlus loggerPlusPlus, LogTableController logTableController, PreferencesController preferencesController){
-        this.loggerPlusPlus = loggerPlusPlus;
+    public GrepperController(LogTableController logTableController, PreferencesController preferencesController){
         this.logTableController = logTableController;
         this.preferences = preferencesController.getPreferences();
         this.listeners = new ArrayList<>();
         this.remainingEntries = new AtomicInteger(0);
         this.grepPanel = new GrepperPanel(this, preferences);
-    }
-
-    public LoggerPlusPlus getLoggerPlusPlus() {
-        return loggerPlusPlus;
     }
 
     public LogTableController getLogTableController() {
@@ -64,15 +59,15 @@ public class GrepperController {
         }
     }
 
-    public IHttpRequestResponseWithMarkers addMarkers(IHttpRequestResponse requestResponse, List<GrepResults.Match> matches) {
-        List<int[]> requestMarkers = new ArrayList<>(), responseMarkers = new ArrayList<>();
+    public HttpRequestResponse addMarkers(HttpRequestResponse requestResponse, List<GrepResults.Match> matches) {
+        List<Marker> requestMarkers = new ArrayList<>(), responseMarkers = new ArrayList<>();
         for (GrepResults.Match match : matches) {
-            int[] marker = new int[]{match.startIndex, match.endIndex};
+            Marker marker = Marker.marker(match.startIndex, match.endIndex);
             if (match.isRequest) requestMarkers.add(marker);
             else responseMarkers.add(marker);
         }
 
-        return LoggerPlusPlus.callbacks.applyMarkers(requestResponse, requestMarkers, responseMarkers);
+        return requestResponse.withRequestMarkers(requestMarkers).withResponseMarkers(responseMarkers);
     }
 
     public void beginSearch(final Pattern pattern, final boolean inScopeOnly, final boolean searchRequests, final boolean searchResponses) {
@@ -80,7 +75,7 @@ public class GrepperController {
         this.searchExecutor = Executors.newFixedThreadPool(searchThreads, new NamedThreadFactory("LPP-Grepper"));
 
         new Thread(() -> {
-            ArrayList<LogEntry> logEntries = new ArrayList<>(loggerPlusPlus.getLogViewController().getLogTableController().getLogTableModel().getData());
+            ArrayList<LogEntry> logEntries = new ArrayList<>(LoggerPlusPlus.instance.getLogViewController().getLogTableController().getLogTableModel().getData());
             remainingEntries.getAndSet(logEntries.size());
 
             this.listeners.forEach(listener -> {
@@ -95,13 +90,13 @@ public class GrepperController {
 
     private GrepResults processEntry(LogEntry entry, Pattern pattern, final boolean searchRequests, final boolean searchResponses) {
         GrepResults grepResults = null;
-        if (entry.requestResponse != null) {
+        if (entry != null) {
             grepResults = new GrepResults(entry);
-            if (entry.requestResponse.getRequest() != null && searchRequests) {
-                processMatches(grepResults, pattern, entry.requestResponse.getRequest(), true);
+            if (entry.getRequestBytes() != null && searchRequests) {
+                processMatches(grepResults, pattern, entry.getRequestBytes(), true);
             }
-            if (entry.requestResponse.getResponse() != null && searchResponses) {
-                processMatches(grepResults, pattern, entry.requestResponse.getResponse(), false);
+            if (entry.getResponseBytes() != null && searchResponses) {
+                processMatches(grepResults, pattern, entry.getResponseBytes(), false);
             }
         }
         return grepResults;
@@ -126,25 +121,28 @@ public class GrepperController {
     private Runnable createProcessThread(final LogEntry logEntry, final Pattern pattern,
                                          final boolean inScopeOnly, final boolean searchRequests, final boolean searchResponses) {
         return () -> {
-            if (Thread.currentThread().isInterrupted()) return;
-            GrepResults grepResults = null;
-            if (!inScopeOnly || LoggerPlusPlus.callbacks.isInScope(logEntry.url)) {
-                grepResults = processEntry(logEntry, pattern, searchRequests, searchResponses);
-            }
-            for (GrepperListener listener : this.listeners) {
-                try {
-                    listener.onEntryProcessed(grepResults);
-                } catch (Exception e) {
-                    e.printStackTrace();
+            try {
+                if (Thread.currentThread().isInterrupted()) return;
+                GrepResults grepResults = null;
+                if (!inScopeOnly || LoggerPlusPlus.isUrlInScope(logEntry.getUrlString())) {
+                    grepResults = processEntry(logEntry, pattern, searchRequests, searchResponses);
                 }
-            }
-            int remaining = remainingEntries.decrementAndGet();
-            if(remaining == 0){
-                for (GrepperListener listener : listeners) {
+                for (GrepperListener listener : this.listeners) {
                     try {
-                        listener.onSearchComplete();
-                    }catch (Exception e){
+                        listener.onEntryProcessed(grepResults);
+                    } catch (Exception e) {
                         e.printStackTrace();
+                    }
+                }
+            }finally {
+                int remaining = remainingEntries.decrementAndGet();
+                if (remaining == 0) {
+                    for (GrepperListener listener : listeners) {
+                        try {
+                            listener.onSearchComplete();
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                        }
                     }
                 }
             }
