@@ -1,10 +1,10 @@
 package com.nccgroup.loggerplusplus.grepper;
 
-import burp.IHttpRequestResponse;
-import burp.IHttpRequestResponseWithMarkers;
 import com.coreyd97.BurpExtenderUtilities.HistoryField;
 import com.coreyd97.BurpExtenderUtilities.PanelBuilder;
 import com.coreyd97.BurpExtenderUtilities.Preferences;
+import com.nccgroup.loggerplusplus.LoggerPlusPlus;
+import com.nccgroup.loggerplusplus.logentry.LogEntry;
 import com.nccgroup.loggerplusplus.logview.entryviewer.RequestViewerController;
 import com.nccgroup.loggerplusplus.util.Globals;
 
@@ -73,7 +73,7 @@ public class GrepperPanel extends JPanel implements GrepperListener {
         });
 
         this.grepResultsTable = new GrepResultsTable(controller);
-        this.requestViewerController = new RequestViewerController(preferences, false, false);
+        this.requestViewerController = new RequestViewerController(preferences);
 
         grepResultsTable.addTreeSelectionListener(treeSelectionEvent -> {
             TreePath selectedPath = treeSelectionEvent.getPath();
@@ -83,7 +83,7 @@ public class GrepperPanel extends JPanel implements GrepperListener {
                 selectedMatch = (GrepResults.Match) selectedPath.getPath()[2];
             }
 
-            IHttpRequestResponse requestResponse = grepResultEntry.getLogEntry().getRequestResponse();
+            LogEntry requestResponse = grepResultEntry.getLogEntry();
             List<GrepResults.Match> matches;
 
             if (selectedMatch != null) {
@@ -92,10 +92,10 @@ public class GrepperPanel extends JPanel implements GrepperListener {
                 matches = grepResultEntry.getMatches();
             }
 
-            IHttpRequestResponseWithMarkers markedRequestResponse = controller.addMarkers(requestResponse, matches);
-            requestViewerController.setDisplayedEntity(markedRequestResponse);
+            requestViewerController.setDisplayedEntity(requestResponse);
 
             //TODO Setup message editor to support highlighting. Code is ready, waiting on API support.
+//            IHttpRequestResponseWithMarkers markedRequestResponse = controller.addMarkers(requestResponse, matches);
             //https://forum.portswigger.net/thread/eeditor-custom-highlighting-991b1a7e?CategoryId=burp-extensions
         });
 
@@ -138,7 +138,7 @@ public class GrepperPanel extends JPanel implements GrepperListener {
         try {
             pattern = Pattern.compile(patternString, Pattern.CASE_INSENSITIVE);
         } catch (PatternSyntaxException e) {
-            JOptionPane.showMessageDialog(JOptionPane.getFrameForComponent(controller.getLoggerPlusPlus().getMainViewController().getUiComponent()), "Pattern Syntax Invalid", "Invalid Pattern", JOptionPane.ERROR_MESSAGE);
+            JOptionPane.showMessageDialog(JOptionPane.getFrameForComponent(LoggerPlusPlus.instance.getMainViewController().getUiComponent()), "Pattern Syntax Invalid", "Invalid Pattern", JOptionPane.ERROR_MESSAGE);
             return;
         }
 
