@@ -1,5 +1,5 @@
 package com.nccgroup.loggerplusplus.logentry;
 
 public enum Status {
-    UNPROCESSED, AWAITING_RESPONSE, PROCESSED, IGNORED
+    UNPROCESSED, AWAITING_RESPONSE, PROCESSED
 }
