@@ -1,60 +1,32 @@
 package com.nccgroup.loggerplusplus.logging;
 
-import burp.IBurpExtenderCallbacks;
+import burp.api.montoya.MontoyaApi;
 import com.coreyd97.BurpExtenderUtilities.IGsonProvider;
-import com.nccgroup.loggerplusplus.LoggerPlusPlus;
 import com.nccgroup.loggerplusplus.util.Globals;
+import lombok.extern.log4j.Log4j2;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.core.Appender;
-import org.apache.logging.log4j.core.LogEvent;
 import org.apache.logging.log4j.core.LoggerContext;
-import org.apache.logging.log4j.core.appender.AbstractAppender;
-import org.apache.logging.log4j.core.config.Configuration;
-import org.apache.logging.log4j.core.layout.PatternLayout;
+import org.apache.logging.log4j.core.appender.ConsoleAppender;
 
+@Log4j2
 public class LoggingController {
 
     private final IGsonProvider gsonProvider;
-    private IBurpExtenderCallbacks callbacks;
     private Level logLevel;
 
-    public LoggingController(IGsonProvider gsonProvider) {
+    public LoggingController(IGsonProvider gsonProvider, MontoyaApi montoyaApi) {
         this.gsonProvider = gsonProvider;
-        this.callbacks = LoggerPlusPlus.callbacks;
-        configureLogger();
-    }
-
-    private void configureLogger() {
-        logLevel = gsonProvider.getGson().fromJson(callbacks.loadExtensionSetting(Globals.PREF_LOG_LEVEL), Level.class);
-
-        if (logLevel == null) { //Handle change from debug boolean to log level.
-            logLevel = Level.INFO;
-            callbacks.saveExtensionSetting(Globals.PREF_LOG_LEVEL, gsonProvider.getGson().toJson(logLevel));
+        logLevel = gsonProvider.getGson().fromJson(montoyaApi.persistence().preferences().getString(Globals.PREF_LOG_LEVEL), Level.class);
+        if(montoyaApi.extension().filename() == null){ //Loaded from classpath. Log to console!
+            LoggerContext context = (LoggerContext) LogManager.getContext(false);
+            ConsoleAppender.Builder appenderBuilder = new ConsoleAppender.Builder();
+            appenderBuilder.setName("ConsoleAppender");
+            ConsoleAppender consoleAppender = appenderBuilder.build();
+            consoleAppender.start();
+            context.getRootLogger().addAppender(consoleAppender);
         }
-
-        LoggerContext context = (LoggerContext) LogManager.getContext(false);
-        Configuration config = context.getConfiguration();
-        PatternLayout logLayout = PatternLayout.newBuilder()
-                .withConfiguration(config)
-                .withPattern("[%-5level] %d{yyyy-MM-dd HH:mm:ss} %msg%n")
-                .build();
-
-        Appender burpAppender = new AbstractAppender("Burp Appender", null, logLayout, false, null) {
-            @Override
-            public void append(LogEvent event) {
-                String message = new String(this.getLayout().toByteArray(event));
-                if (event.getLevel().isMoreSpecificThan(Level.ERROR)) {
-                    callbacks.printError(message);
-                } else {
-                    callbacks.printOutput(message);
-                }
-            }
-        };
-        burpAppender.start();
-
-        context.getConfiguration().getRootLogger().addAppender(burpAppender, Level.INFO, null);
-        context.updateLoggers();
+        setLogLevel(logLevel);
     }
 
     public void setLogLevel(Level logLevel) {
