@@ -1,28 +1,29 @@
 package com.nccgroup.loggerplusplus.logview.processor;
 
-import burp.IHttpRequestResponse;
+import burp.api.montoya.core.Annotations;
 import com.nccgroup.loggerplusplus.util.Globals;
+import org.apache.commons.lang3.StringUtils;
 
-import java.util.UUID;
 import java.util.regex.Matcher;
 
 public class LogProcessorHelper {
 
-    public static void tagRequestResponseWithUUID(String instanceIdentifier, UUID uuid, IHttpRequestResponse requestResponse){
-        String originalComment = requestResponse.getComment() != null ? requestResponse.getComment() : "";
-        requestResponse.setComment(originalComment + "$LPP:" + instanceIdentifier + ":" + uuid + "$");
+    public static Annotations addIdentifierInComment(Integer identifier, Annotations annotations) {
+        String originalComment = annotations.notes() != null ? annotations.notes() : "";
+        annotations = annotations.withNotes(originalComment + "$LPP:" + identifier + "$");
+        return annotations;
     }
 
-    public static UUID extractAndRemoveUUIDFromRequestResponseComment(String instanceIdentifier, IHttpRequestResponse requestResponse){
-        UUID uuid = null;
-        if(requestResponse.getComment() != null) {
-            Matcher matcher = Globals.UUID_COMMENT_PATTERN.matcher(requestResponse.getComment());
-            if (matcher.find() && matcher.group(1).equals(instanceIdentifier)) {
-                uuid = UUID.fromString(matcher.group(2));
-                requestResponse.setComment(matcher.replaceAll(""));
+    public static Object[] extractAndRemoveIdentifierFromRequestResponseComment(Annotations annotations) {
+        Integer identifier = null;
+        if (!StringUtils.isEmpty(annotations.notes())) {
+            Matcher matcher = Globals.LOG_ENTRY_ID_PATTERN.matcher(annotations.notes());
+            if (matcher.find()) {
+                identifier = Integer.parseInt(matcher.group(1));
+                annotations = annotations.withNotes(matcher.replaceAll(""));
             }
         }
 
-        return uuid;
+        return new Object[]{identifier,annotations};
     }
 }
