@@ -3,7 +3,6 @@ package com.nccgroup.loggerplusplus.util;
 import com.coreyd97.BurpExtenderUtilities.Preferences;
 import com.nccgroup.loggerplusplus.LoggerPlusPlus;
 import com.nccgroup.loggerplusplus.logentry.LogEntryField;
-import org.apache.commons.codec.digest.DigestUtils;
 
 import javax.swing.*;
 import javax.swing.filechooser.FileNameExtensionFilter;
@@ -11,7 +10,6 @@ import javax.swing.text.html.HTMLEditorKit;
 import java.awt.*;
 import java.io.File;
 import java.io.UnsupportedEncodingException;
-import java.security.MessageDigest;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -92,7 +90,7 @@ public class MoreHelp {
 		final Object[] options = msgOptions;
 		final int[] choice = new int[1];
 		choice[0] = 0;
-		choice[0] = JOptionPane.showOptionDialog(null, strQuestion, strTitle, JOptionPane.YES_NO_CANCEL_OPTION,
+		choice[0] = JOptionPane.showOptionDialog(LoggerPlusPlus.instance.getLoggerFrame(), strQuestion, strTitle, JOptionPane.YES_NO_CANCEL_OPTION,
 				JOptionPane.QUESTION_MESSAGE, null, options, options[0]);
 		return choice[0];
 	}
@@ -112,7 +110,7 @@ public class MoreHelp {
 
 	// public static void checkForUpdate(boolean showMessages) {
 	// new Thread(() -> {
-	// IExtensionHelpers helper = LoggerPlusPlus.callbacks.getHelpers();
+	// IExtensionHelpers helper = LoggerPlusPlus.montoya.getHelpers();
 	// Double currenVersion = Globals.VERSION;
 	// Double latestVersion = 0.0;
 	// int updateStatus = -1;
@@ -122,7 +120,7 @@ public class MoreHelp {
 	// URL changeLogURL = new URL(Globals.CHANGELOG);
 	// byte[] request = helper.buildHttpRequest(changeLogURL);
 	// byte[] response =
-	// LoggerPlusPlus.callbacks.makeHttpRequest(changeLogURL.getHost(), 443, true,
+	// LoggerPlusPlus.montoya.makeHttpRequest(changeLogURL.getHost(), 443, true,
 	// request);
 	//
 	// if (response != null) {
@@ -150,27 +148,27 @@ public class MoreHelp {
 	//
 	// }
 	// } catch (Exception e) {
-	// LoggerPlusPlus.callbacks.printError(e.getMessage());
+	// LoggerPlusPlus.montoya.printError(e.getMessage());
 	// }
 	//
 	// switch (updateStatus) {
 	// case -1:
 	// updateMessage = "Check for update failed: Could not get a proper response
 	// from " + Globals.CHANGELOG;
-	// LoggerPlusPlus.callbacks.printError(updateMessage);
+	// LoggerPlusPlus.montoya.printError(updateMessage);
 	// break;
 	// case 0:
 	// updateMessage = "This version is up to date.";
-	// LoggerPlusPlus.callbacks.printOutput(updateMessage);
+	// LoggerPlusPlus.montoya.printOutput(updateMessage);
 	// break;
 	// case 1:
 	// updateMessage = "Version " + latestVersion.toString() + " is available via
 	// GitHub. Visit the extension homepage.";
-	// if (LoggerPlusPlus.callbacks.isExtensionBapp()) {
+	// if (LoggerPlusPlus.montoya.isExtensionBapp()) {
 	// updateMessage += "\nAs you are using BApp Store, you have to remove it first
 	// and download the Jar file from the GitHub repository. ";
 	// } else {
-	// if (LoggerPlusPlus.callbacks.getExtensionFilename() != null) {
+	// if (LoggerPlusPlus.montoya.getExtensionFilename() != null) {
 	// int res = MoreHelp.askConfirmMessage("Update Available", "An update is
 	// available. Would you like to update now?", new String[]{"Yes", "No"});
 	// if (res == JOptionPane.OK_OPTION) {
@@ -178,7 +176,7 @@ public class MoreHelp {
 	// //TODO FIXME
 	// URL updateUrl = new URL(Globals.UPDATE_URL);
 	// InputStream input = updateUrl.openStream();
-	// Path outputPath = Paths.get(LoggerPlusPlus.callbacks.getExtensionFilename());
+	// Path outputPath = Paths.get(LoggerPlusPlus.montoya.getExtensionFilename());
 	// Files.copy(input, outputPath, StandardCopyOption.REPLACE_EXISTING);
 	// } catch (Exception e) {
 	// MoreHelp.showMessage("Could not update the plugin. Please visit the extension
@@ -187,17 +185,17 @@ public class MoreHelp {
 	// }
 	// MoreHelp.showMessage("Update complete. Re-enable the plugin in the extensions
 	// tab to continue.");
-	// LoggerPlusPlus.callbacks.unloadExtension();
+	// LoggerPlusPlus.montoya.unloadExtension();
 	// return;
 	// }
 	// }
 	// }
-	// LoggerPlusPlus.callbacks.printOutput(updateMessage);
+	// LoggerPlusPlus.montoya.printOutput(updateMessage);
 	// break;
 	// case 2:
 	// updateMessage = "This version is more up to date than the GitHub version! Are
 	// you a time traveler? or just a keen ninja? ;)";
-	// LoggerPlusPlus.callbacks.printOutput(updateMessage);
+	// LoggerPlusPlus.montoya.printOutput(updateMessage);
 	// break;
 	// }
 	// if (!showMessages) return;
